build : detailView -> detail 네이밍 변경
